refactor(netutil): Consolidate HTTP client and slice utility functions
