+++ internal/fetcher/http_headers.go
@@ -1,6 +1,8 @@
 package fetcher
 
+import "github.com/wallentx/jobscout/internal/netutil"
+
 const (
-	jobscoutUserAgent    = "jobscout (+https://github.com/wallentx/jobscout)"
-	browserLikeUserAgent = "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/125.0 Safari/537.36 jobscout (+https://github.com/wallentx/jobscout)"
+	jobscoutUserAgent    = netutil.JobscoutUserAgent
+	browserLikeUserAgent = netutil.BrowserLikeUserAgent
 )
